DONE: -[X] Changes in layouts -[X] Changes in Dataclasses -[X] Added Subtotal and Taxes in BillFragment.kt -[X] Responsive layout -[X] Added View-Invoice screen

+++ app/src/main/java/com/codebyzebru/myapplication/dataclasses/AddInventoryDataClass.kt
@@ -4,6 +4,6 @@ class AddInventoryDataClass(
     val key: String,
     val productName: String,
     val purchasingPrice: String?,
-    val sellingPrice: Int,
-    val productQty: Int
+    val sellingPrice: Float,
+    val productQty: Float
 )
